+++ lib/Extension/LanguageServerReferenceFinder/LanguageServerReferenceFinderExtension.php
@@ -40,7 +40,8 @@ class LanguageServerReferenceFinderExtension implements Extension
             return new TypeDefinitionHandler(
                 $container->get(LanguageServerExtension::SERVICE_SESSION_WORKSPACE),
                 $container->get(ReferenceFinderExtension::SERVICE_TYPE_LOCATOR),
-                $container->get(LocationConverter::class)
+                $container->get(LocationConverter::class),
+                $container->get(ClientApi::class),
             );
         }, [ LanguageServerExtension::TAG_METHOD_HANDLER => [] ]);